Add ClassMemberInfo as super class of FieldInfo and MethodInfo
